+++ toolbox/+igraph/private/mustBePartition.m
@@ -11,9 +11,13 @@ function mustBePartition(membership)
         msg = "Membership vector must be real";
     end
 
-    if ~(isvector(membership))
-        msg = "Membership vector must be a vector";
-    end
+    % TODO: Reconsider, should be able to handle multiple partitions in a
+    % single matrix such that each row is a partition. This is useful for
+    % partitions that are closely related such as hierarchical community
+    % structure.
+    % if ~(isvector(membership))
+    %     msg = "Membership vector must be a vector";
+    % end
 
     if ~strcmp(msg, "")
         msg = "Value must be a membership vector\n\n" + msg;