+++ toolbox/+igraph/reindexMembership.m
@@ -1,0 +1,7 @@
+function memb = reindexMembership(memb)
+    arguments
+        memb {mustBePartition};
+    end
+
+    memb = mexIgraphDispatcher(mfilename(), memb);
+end